Deprecated: `GraphQL\Schema` moved to `GraphQL\Type\Schema`
