+++ src/config/config.php
@@ -100,7 +100,7 @@ return [
     /*
      * The schemas for query and/or mutation. It expects an array to provide
      * both the 'query' fields and the 'mutation' fields. You can also
-     * provide an GraphQL\Schema object directly.
+     * provide an GraphQL\Type\Schema object directly.
      *
      * Example:
      *
